🚚 move: QUOTES -> constants, getDailyProgrammingQuote -> utils
