+++ app/src/Home/constants/QUOTES.ts
